requirements.txt
asn1crypto==0.24.0
backports-abc==0.5
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
cryptography==2.5
fabric==2.4.0
idna==2.8
invoke==1.2.0
numpy==1.16.1
paramiko==2.4.2
Pillow==5.4.1
psycopg2==2.7.7
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
python-dotenv==0.10.1
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
tornado==5.1.1
urllib3==1.24.1
